refactor(utils): extract leftover helpers into mtui/misc.py
Move the three remaining helpers (requires_update decorator,
DictWithInjections container, SUTParse argv-format helper) verbatim
out of mtui/utils.py into a new mtui/misc.py module. None of them fit
the four named target modules (colors/completion/fileops/term) so they
land in a clearly-labelled holding pen rather than overloading one of
the topical modules.

utils.py is now a pure re-export shim. The next commit rewires all
in-tree call sites to import from the new homes and deletes utils.py
entirely.

Fifth commit of the Phase 5b / C8 split (PLAN.md item 11).
